GateクラスとTicketクラスを別ファイルに分けた。
